requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
meshio==5.3.5
numpy==2.1.1
pillow==10.4.0
Pygments==2.18.0
rich==13.8.1
trimesh==4.4.9
